chore(ai): the validating-proxy machinery moves below both barrels (#16710)
